+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # v 0.2.9
 
 * Dragonborn breath weapon was not parsing DC save type correctly.
+* Better detection of optional choices like Mask of the Wild for variant Half Elf's.
 
 # v 0.2.7/0.2.8