+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * Better detection of optional choices like Mask of the Wild for variant Half Elf's.
 * Support Dual Wielding AC Bonus.
 * Better UI messages for homebrew.
+* Add ID tags to most items and feats for future [redacted].
 
 # v 0.2.7/0.2.8